auth_service/requirement.txt
Flask==2.2.5
PyJWT==2.8.0
requests==2.31.0